+++ api/proto/main.proto
@@ -7,7 +7,7 @@ service AIService {
 }
 
 message AskRequest {
-  string id = 1;
+  int64 id = 1;
   string text = 2;
 }
